+++ src/search/pruning/stubborn_sets_atom_centric.cc
@@ -3,9 +3,7 @@
 #include "../option_parser.h"
 #include "../plugin.h"
 
-#include "../utils/logging.h"
 #include "../utils/markup.h"
-#include "../utils/memory.h"
 
 using namespace std;
 
@@ -195,7 +193,7 @@ void StubbornSetsAtomCentric::enqueue_interferers(int op) {
     }
 }
 
-void StubbornSetsAtomCentric::initialize_stubborn_set(const State &state) {
+void StubbornSetsAtomCentric::compute_stubborn_set(const State &state) {
     assert(producer_queue.empty());
     assert(consumer_queue.empty());
     // Reset data structures from previous call.